refactor:调整餐饮-购物车js文件
