+++ api/env.js
@@ -1,5 +1,5 @@
-// export const baseUrl = 'http://localhost:18000'
-export const baseUrl = 'https://api.itwray.com'
+export const baseUrl = 'http://localhost:18000'
+// export const baseUrl = 'https://api.itwray.com'
 
 export const token_key = 'iwtoken'
